split mock function
