+++ routes/__tests__/dhb.test.js
@@ -1,17 +1,7 @@
 const request = require("supertest");
-const dhbRouter = require("../dhb");
 const express = require("express");
-
-const generateMockData = (length) => {
-  let mockData = [];
-  for (let i = 0; i < length; i++) {
-    mockData.push({
-      value: i + 1,
-      display: `${Math.random().toString(36).substring(2, 11)}`,
-    });
-  }
-  return mockData;
-};
+const dhbRouter = require("../dhb");
+const generateMockData = require("./mockFunction");
 
 const getNorthDhbs = jest.fn();
 const getSouthDhbs = jest.fn();
